+++ src/shared/dto/AppState.ts
@@ -1,0 +1,8 @@
+enum AppState {
+  init,
+  loadingConfigs,
+  readingHistory,
+  ready,
+}
+
+export default AppState;
